refactor: tiedtke types.py split + lohmann_2m package split (pure moves)
Maintainability review C.4: the four Tiedtke data structs move to
tiedtke_nordeng/types.py (re-exported, no import cycles); the 3700-line
lohmann_2m.py becomes the lohmann_2m/ package (types, sedimentation_melt,
deposition_freezing, precip, assembly, scheme) split strictly on
function boundaries with all 18 public names re-exported.

vmap hoisting (perf plan Step 2) NOT included: the plan's literal step
already landed as ComposablePhysics(vectorize_columns=True); the
remaining per-term cleanup means rewriting three genuinely
single-column schemes broadcasting-native (the 2M carries level-to-
level flux carries through lax.scan) — deep surgery deferred to a
dedicated PR with per-term parity pins.

Validation: full fast suite identical before/after (1244 passed, 17
skipped), ruff clean, zero pin changes.
